INT-1723 Neutron to create Elasticsearch Alias for people-summary index (#101)
* INT-1723 Neutron to create Elasticsearch Alias for people-summary index

* INT-1723 Neutron to create Elasticsearch Alias for people-summary index

+++ src/main/java/gov/ca/cwds/neutron/rocket/IndexResetRocket.java
@@ -1,5 +1,6 @@
 package gov.ca.cwds.neutron.rocket;
 
+import java.text.SimpleDateFormat;
 import java.util.Date;
 
 import org.apache.commons.lang3.StringUtils;
@@ -52,15 +53,28 @@ public abstract class IndexResetRocket
     try {
       // If index name is provided, use it, else take alias from ES config.
       final String indexNameOverride = getFlightPlan().getIndexName();
-      final String effectiveIndexName =
+      String effectiveIndexName =
           StringUtils.isBlank(indexNameOverride) ? esDao.getConfig().getElasticsearchAlias()
               : indexNameOverride;
-      getFlightPlan().setIndexName(effectiveIndexName);
 
-      // Drop index first, if requested.
-      if (getFlightPlan().isDropIndex()) {
-        esDao.deleteIndex(effectiveIndexName);
+      if (!getFlightPlan().isLastRunMode() ) {
+        //Force new index name for Initial Load when Index name not provided
+        if (StringUtils.isBlank(indexNameOverride)) {
+          effectiveIndexName = effectiveIndexName.concat("_")
+              .concat(new SimpleDateFormat("yyyy.MM.dd.HH.mm.ss").format(new Date()));
+        }
+        // Drop index first, if requested and name provided
+        if (getFlightPlan().isDropIndex() && !StringUtils.isBlank(indexNameOverride)) {
+          esDao.deleteIndex(effectiveIndexName);
+        }
+      } else {
+        // Drop index first, if requested
+        if (getFlightPlan().isDropIndex()) {
+          esDao.deleteIndex(effectiveIndexName);
+        }
       }
+
+      getFlightPlan().setIndexName(effectiveIndexName);
 
       // If the index is missing, create it.
       final String documentType = esDao.getConfig().getElasticsearchDocType();
